LC-factorial-trailing-zeroes [Runtime: N/A; Memory: N/A]

+++ math/factorial-trailing-zeroes.cpp
@@ -1,13 +1,13 @@
 class Solution {
 public:
-    int fact(int n){
+    long long fact(int n){
         if(n==0 || n==1){
             return 1;
         }
         return n*fact(n-1);
     }
     int trailingZeroes(int n) {
-        int fa=fact(n);
+        long long fa=fact(n);
         if(fa<10) return 0;
         int count=0;
         while(fa%10==0){
